Create Continents.java
LEARN JAVA
Continents and Cities
Let's practice the Java switch statement that you learned about.

In this project, you will write a program that will print out a continent and the largest city in that continent, based on the value of an integer.

The instructions provided are general guidelines. Upon completion of the project, feel free to add functionality of your own.